requirements.txt
Flask
opencv-python
numpy
matplotlib
Pillow
pafy
youtube_dl
tensorflow
streamlit